Move PipelineQueue into its own file
