+++ Pipeline/Pipeline.swift
@@ -8,61 +8,6 @@
 
 import Foundation
 
-public class PipelineQueue {
-    public enum QueueLevel {
-        case Main
-        case QOS(NSQualityOfService)
-    }
-
-    private let internalQueue: NSOperationQueue
-    private var allOperations: [NSOperation] = []
-
-    public let queueLevel: QueueLevel
-    public var suspended = true {
-        didSet {
-            internalQueue.suspended = suspended
-        }
-    }
-    public var operations: [NSOperation] {
-        return allOperations.filter { !$0.finished }
-    }
-
-    public init(_ queueLevel: QueueLevel = .QOS(.Default)) {
-        self.queueLevel = queueLevel
-        switch queueLevel {
-        case .Main:
-            internalQueue = NSOperationQueue.mainQueue()
-        case .QOS(let QOS):
-            internalQueue = NSOperationQueue()
-            internalQueue.qualityOfService = QOS
-        }
-    }
-
-    public func cancelAllOperations() {
-        for operation in allOperations {
-            operation.cancel()
-        }
-        allOperations = []
-    }
-
-    public func addOperation<Operation where Operation: NSOperation, Operation: Pipelinable>(operation: Operation, _ queue: QueueLevel? = nil) {
-        allOperations.append(operation)
-
-        if let queue = queue {
-            switch queue {
-            case .Main:
-                PipelineQueue(.Main).addOperation(operation)
-            case .QOS(let QOS):
-                operation.qualityOfService = QOS
-                internalQueue.addOperation(operation)
-            }
-        }
-        else {
-            internalQueue.addOperation(operation)
-        }
-    }
-}
-
 public class Pipeline {
     internal enum State {
         case Ready
